+++ src/main.rs
@@ -22,7 +22,7 @@ pub struct Config {
 
 	/// the length of the sequence to output, i.e.:
 	/// 0, 1, 1, 2, 3, 5 would be the result of -c 6
-	pub count: u32,
+	pub count: u128,
 
 	/// Print on one line
 	pub one_line: bool,
